safe_bite/backend/requirements.txt
Flask
gTTS
google-genai
python-dotenv
datetime